feat: streamextractor and streamloader
