+++ src/atc/delta/__init__.py
@@ -1,4 +1,3 @@
 from .autoloaderstream_handle import AutoloaderStreamHandle  # noqa: F401
 from .db_handle import DbHandle  # noqa: F401
 from .delta_handle import DeltaHandle  # noqa: F401
-from .deltastream_handle import DeltaStreamHandle  # noqa: F401
