+++ src/atc/delta/delta_handle.py
@@ -1,21 +1,61 @@
-from typing import List, Union
+from datetime import datetime
+from typing import List, Optional, Union
 
 from pyspark.sql import DataFrame
 
 from atc.configurator.configurator import Configurator
+from atc.exceptions import AtcException
 from atc.functions import get_unique_tempview_name, init_dbutils
 from atc.spark import Spark
-from atc.tables.SparkHandle import SparkHandle
+from atc.tables import TableHandle
 from atc.utils.CheckDfMerge import CheckDfMerge
 from atc.utils.GetMergeStatement import GetMergeStatement
 
 
-class DeltaHandle(SparkHandle):
-    def __init__(self, name: str, location: str = None, data_format: str = "delta"):
+class DeltaHandleException(AtcException):
+    pass
 
-        super().__init__(name, location, data_format)
 
+class DeltaHandleInvalidName(DeltaHandleException):
+    pass
+
+
+class DeltaHandleInvalidFormat(DeltaHandleException):
+    pass
+
+
+class DeltaHandle(TableHandle):
+    def __init__(
+        self,
+        name: str,
+        location: str = None,
+        data_format: str = "delta",
+        options_dict: dict = None,
+        ignore_changes: bool = True,
+        stream_start: datetime = None,
+        max_bytes_per_trigger: int = None,
+        # checkpoint_path: str = None,
+    ):
+        """ """
+        self._name = name
+        self._location = location
+        self._data_format = data_format
+
+        self._location = location
+
+        self._partitioning: Optional[List[str]] = None
         self._validate()
+
+        self._options_dict = options_dict
+        self._options_dict["ignoreChanges"] = ignore_changes
+
+        if stream_start:
+            self._options_dict["startingTimestamp"] = stream_start.strftime(
+                "%Y-%m-%dT%H:%M:%S.%fZ"
+            )
+
+        if max_bytes_per_trigger:
+            self._options_dict["maxBytesPerTrigger"] = max_bytes_per_trigger
 
     @classmethod
     def from_tc(cls, id: str) -> "DeltaHandle":
@@ -55,8 +95,11 @@ class DeltaHandle(SparkHandle):
     def truncate(self) -> None:
         Spark.get().sql(f"TRUNCATE TABLE {self._name};")
 
+        self.remove_checkpoint()
+
     def drop(self) -> None:
         Spark.get().sql(f"DROP TABLE IF EXISTS {self._name};")
+        self.remove_checkpoint()
 
     def drop_and_delete(self) -> None:
         self.drop()
@@ -68,7 +111,6 @@ class DeltaHandle(SparkHandle):
         df: DataFrame,
         join_cols: List[str],
     ) -> Union[DataFrame, None]:
-
         if df is None:
             return None
 
@@ -121,3 +163,96 @@ class DeltaHandle(SparkHandle):
         print("Incremental Base - incremental load with merge")
 
         return df
+
+    def read_stream(self) -> DataFrame:
+        assert (
+            Spark.version() >= Spark.DATABRICKS_RUNTIME_10_4
+        ), f"Streaming not available for Spark version {Spark.version()}"
+
+        reader = (
+            Spark.get().readStream.format(self._format).options(**self._options_dict)
+        )
+        if self._location:
+            df = reader.load(self._location)
+        else:
+            df = reader.table(self._table_name)
+
+        return df
+
+    def _validate(self):
+        """Validates that the name is either db.table or just table."""
+        if not self._name:
+            if not self._location:
+                raise DeltaHandleInvalidName(
+                    "Cannot create DeltaHandle without name or path"
+                )
+            self._name = f"delta.`{self._location}`"
+        else:
+            name_parts = self._name.split(".")
+            if len(name_parts) == 1:
+                self._db = None
+                self._table_name = name_parts[0]
+            elif len(name_parts) == 2:
+                self._db = name_parts[0]
+                self._table_name = name_parts[1]
+            else:
+                raise DeltaHandleInvalidName(f"Could not parse name {self._name}")
+
+        # only format delta is supported.
+        if self._data_format != "delta":
+            raise DeltaHandleInvalidFormat("Only format delta is supported.")
+
+    def create_hive_table(self) -> None:
+        self.remove_checkpoint()
+        sql = f"CREATE TABLE IF NOT EXISTS {self._name} "
+        if self._location:
+            sql += f" USING DELTA LOCATION '{self._location}'"
+        Spark.get().sql(sql)
+
+    def recreate_hive_table(self):
+        self.drop()
+        self.create_hive_table()
+
+    def get_partitioning(self):
+        """The result of DESCRIBE TABLE tablename is like this:
+        +-----------------+---------------+-------+
+        |         col_name|      data_type|comment|
+        +-----------------+---------------+-------+
+        |           mycolA|         string|       |
+        |           myColB|            int|       |
+        |                 |               |       |
+        |   # Partitioning|               |       |
+        |           Part 0|         mycolA|       |
+        +-----------------+---------------+-------+
+        but this method return the partitioning in the form ['mycolA'],
+        if there is no partitioning, an empty list is returned.
+        """
+        if self._partitioning is None:
+            # create an iterator object and use it in two steps
+            rows_iter = iter(
+                Spark.get().sql(f"DESCRIBE TABLE {self.get_tablename()}").collect()
+            )
+
+            # roll over the iterator until you see the title line
+            for row in rows_iter:
+                # discard rows until the important section header
+                if row.col_name.strip() == "# Partitioning":
+                    break
+            # at this point, the iterator has moved past the section heading
+            # leaving only the rows with "Part 1" etc.
+
+            # create a list from the rest of the iterator like [(0,colA), (1,colB)]
+            parts = [
+                (int(row.col_name[5:]), row.data_type)
+                for row in rows_iter
+                if row.col_name.startswith("Part ")
+            ]
+            # sort, just in case the parts were out of order.
+            parts.sort()
+
+            # discard the index and put into an ordered list.
+            self._partitioning = [p[1] for p in parts]
+        return self._partitioning
+
+    def get_tablename(self) -> str:
+        return self._name
