+++ src/atc/utils/FileExists.py
@@ -8,5 +8,8 @@ def file_exists(path: str):
     try:
         init_dbutils().fs.ls(path)
         return True
-    except Exception:
-        return False
+    except Exception as e:
+        if "java.io.FileNotFoundException" in str(e):
+            return False
+        else:
+            raise e
